+++ ruptures/detection/_detection/ekcpd_computation.h
@@ -1,0 +1,1 @@
+void ekcpd_compute(double *signal, int n_samples, int n_dims, int n_bkps, int min_size, void *kernelDescObj, int *M_path_res);