+++ ruptures/detection/dynp.py
@@ -28,6 +28,7 @@ class Dynp(BaseEstimator):
         if custom_cost is not None and isinstance(custom_cost, BaseCost):
             self.cost = custom_cost
         else:
+            self.model_name = model
             if params is None:
                 self.cost = cost_factory(model=model)
             else:
@@ -61,7 +62,7 @@ class Dynp(BaseEstimator):
             for bkp in multiple_of_jump:
                 n_samples = bkp - start
                 # first check if left subproblem is possible
-                if sanity_check(n_samples, n_bkps, jump, min_size):
+                if sanity_check(n_samples, n_bkps - 1, jump, min_size):
                     # second check if the right subproblem has enough points
                     if end - bkp >= min_size:
                         admissible_bkps.append(bkp)